chore(docs): cleanup wording and syntax
This makes sure that the heading levels are consistent with the heading
levels expected by Antora, i.e. page needs to start at heading 1 in
order to have the title set.

Makes sure that code block titles are set at the right place.

Rewords the examples, the trying out locally in particular and
structures the examples in sections.